+++ src/oldnews/providers/main.py
@@ -18,6 +18,7 @@ from ..commands import (
     Next,
     NextUnread,
     OpenArticle,
+    OpenOrigin,
     Previous,
     PreviousUnread,
     RefreshFromTheOldReader,
@@ -41,6 +42,7 @@ class MainCommands(CommandsProvider):
         yield from self.maybe(Previous)
         yield from self.maybe(PreviousUnread)
         yield from self.maybe(OpenArticle)
+        yield from self.maybe(OpenOrigin)
         yield from self.maybe(MarkAllRead)
         yield ToggleShowAll()
         yield RefreshFromTheOldReader()